Flytt knappene ut i generell komponent
